requirements.txt
streamlit
pandas
altair
gspread
google-auth
openpyxl
python-docx
pyarrow